Move Watchers story into unit tests
